Replace QVector<float> with QVector<double> in QtQuick items to fix problems with Qt 5.14

+++ qtquick_items/IrregularCircleItem.cpp
@@ -38,7 +38,7 @@ void IrregularCircleItem::setOuterColor(const QColor& color) {
     update();
 }
 
-void IrregularCircleItem::setRadii(const QVector<float>& radii) {
+void IrregularCircleItem::setRadii(const QVector<double>& radii) {
     m_radii = radii;
     emit radiiChanged();
     update();
